requirements.txt
Django==1.8
Markdown==2.6.2
MySQL-python==1.2.5
Pillow==3.0.0
PyJWT==1.4.0
PyRSS2Gen==1.0.0
Twisted==12.2.0
altgraph==0.10.1
bdist-mpkg==0.4.4
beautifulsoup4==4.4.0
bonjour-py==0.3
certifi==2015.9.6.2
cffi==1.3.0
cryptography==1.1
django-allauth==0.23.0
django-contact-form==1.0
django-crispy-forms==1.5.2
django-dajax==0.9.2
django-dajaxice==0.7
django-generic-ratings==0.7
django-haystack==2.4.1
django-jsonview==0.5.0
django-likes==0.2
django-listable==0.3.1
django-markdown==0.8.4
django-secretballot==0.4.0
django-social-auth==0.7.28
django-split-json-widget==1.15
elasticsearch==1.9.0
enum34==1.0.4
facepy==1.0.7
httplib2==0.9.2
idna==2.0
ipaddress==1.0.14
lxml==3.4.4
macholib==1.5
matplotlib==1.1.1
modulegraph==0.10.1
mysql-utilities==1.5.4
ndg-httpsclient==0.4.0
numpy==1.6.2
oauth2==1.9.0.post1
oauthlib==1.0.3
py2app==0.7.1
pyOpenSSL==0.15.1
pyasn1==0.1.9
pycparser==2.14
pyobjc-core==2.3.2a0
pyobjc-framework-AddressBook==2.3.2a0
pyobjc-framework-AppleScriptKit==2.3.2a0
pyobjc-framework-AppleScriptObjC==2.3.2a0
pyobjc-framework-Automator==2.3.2a0
pyobjc-framework-CFNetwork==2.3.2a0
pyobjc-framework-Cocoa==2.3.2a0
pyobjc-framework-Collaboration==2.3.2a0
pyobjc-framework-CoreData==2.3.2a0
pyobjc-framework-CoreLocation==2.3.2a0
pyobjc-framework-CoreText==2.3.2a0
pyobjc-framework-DictionaryServices==2.3.2a0
pyobjc-framework-ExceptionHandling==2.3.2a0
pyobjc-framework-FSEvents==2.3.2a0
pyobjc-framework-InputMethodKit==2.3.2a0
pyobjc-framework-InstallerPlugins==2.3.2a0
pyobjc-framework-InstantMessage==2.3.2a0
pyobjc-framework-InterfaceBuilderKit==2.3.2a0
pyobjc-framework-LatentSemanticMapping==2.3.2a0
pyobjc-framework-LaunchServices==2.3.2a0
pyobjc-framework-Message==2.3.2a0
pyobjc-framework-OpenDirectory==2.3.2a0
pyobjc-framework-PreferencePanes==2.3.2a0
pyobjc-framework-PubSub==2.3.2a0
pyobjc-framework-QTKit==2.3.2a0
pyobjc-framework-Quartz==2.3.2a0
pyobjc-framework-ScreenSaver==2.3.2a0
pyobjc-framework-ScriptingBridge==2.3.2a0
pyobjc-framework-SearchKit==2.3.2a0
pyobjc-framework-ServiceManagement==2.3.2a0
pyobjc-framework-SyncServices==2.3.2a0
pyobjc-framework-SystemConfiguration==2.3.2a0
pyobjc-framework-WebKit==2.3.2a0
pyobjc-framework-XgridFoundation==2.3.2a0
pysolr==3.3.2
python-dateutil==1.5
python-openid==2.2.5
python-social-auth==0.2.13
pytz==2012d
requests==2.8.1
requests-oauthlib==0.5.0
runscript==0.2.2
scipy==0.11.0
simplejson==3.8.1
six==1.10.0
stripe==1.27.0
urllib3==1.12
virtualenv==12.1.1
wsgiref==0.1.2
xattr==0.6.4
zope.interface==3.8.0
